+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Comparing the old and new hash data via JSON Patch, and transforming that into a set of translation change templates
 - Compiling a full set of changes to make linked to each specific destination translation file
 - Logging messages informing the user about what we loaded, what we couldn't load, how many changes we've detected, etc
+- Execute the changeset using a rate-limited scheduler, hitting the Google Cloud Translation API, and setting the new values in the translation data
 
 ## [0.2.0] - 2020-04-26